recipes-bsp/u-boot: Disable saveenv shell command
balenaOS uboot exports specific environment variables
in resinOS_uEnv.txt and does not have any use-cases of
saving the entire uboot environment in any device specific
partitions. This because it relies on the default environment being
stored in the u-boot binary.

Let's disable the saveenv command and avoid potential incorrect
usage which may overwrite the partition table, resin-boot filesystem
or other areas of the eMMC that may be pre-configured by the BSP.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>