resin-init-flasher: remove build time check on INTERNAL_DEVICE_KERNEL
Non-flasher device types are not currently defining this variable that
is required for the installer.

Make the resin-init-flasher script check for it's existance before using
it instead of having a build time check.

Devices that use an installer image will need to define it in their
integration layer.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>