Allow to specify multiple trusted certificates when building for secure boot
At this moment we only allow a single certificate for PK, KEK and db,
this PR makes it possible to specify a list of certificates using

SIGN_EFI_KEY_ID=key1,key2,key3

The key used for signing will always be the first one specified
("key1" in the example above), the rest will just be loaded as trusted.
This is necessary to implement a proper key rotation.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>